fastapi/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
loguru==0.7.2
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.2
passlib==1.7.4
pinecone-client==2.2.4
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
